🚚 [templates] Rename module `{bindvariables => renderbind}`
